backend/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.7
django-cors-headers==3.7.0
django-jet-reboot==1.1.1
django-livereload==1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
google-api-python-client==1.4.1
httplib2==0.19.0
oauth2client==4.1.3
Pillow==8.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.0.1
pyparsing==2.4.7
pytz==2021.1
rsa==4.7.2
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
uritemplate==3.0.1
